% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_data.R
\name{connect_to_data}
\alias{connect_to_data}
\title{Allows you to easily  connect to the database}
\usage{
connect_to_data(username, password)
}
\arguments{
\item{username}{Username as a string}

\item{password}{Password as a string}
}
\value{
Returns an a connect object called sm. Refer to the object as sm. ie assign the results of the function to an object called sm
}
\description{
Takes in the username and password as strings
}
