% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{drop}
\alias{drop}
\title{Easily drop tables from explorys database}
\usage{
drop(sandbox, table)
}
\arguments{
\item{sandbox}{Name of the writable location usually sandbox where cohort is saved.}

\item{table}{Name of the table to be deleted.}
}
\value{
Nothing is returned but the data is deleted from the sandbox
}
\description{
Takes in the name of the sandbox and table as character strings
}
