% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diagnosis.R
\name{get_diagnosis}
\alias{get_diagnosis}
\title{Get a dataset with flags indicating if patient has a diagnosis}
\usage{
get_diagnosis(
  supermart,
  sandbox,
  cohort,
  dx_codes_df,
  dx_codes_col,
  grouping_col,
  index_col,
  date_condition
)
}
\arguments{
\item{supermart}{Name of the supermart where the data is stored. Eg. supermart_392}

\item{sandbox}{Name of the writable location usually sandbox where cohort is saved}

\item{cohort}{Name of the cohort with explorys_patient_id's.}
}
\description{
This function provides a dataframe with flags indicating if a patient has a diagnosis or not based on icd9, icd10 and snomed codes.
The flags are 1 whichindicate patient had the medication or 0 which indicates patient did not. A missing flag means that
the patient did not get any diagnosis at all in the time frame specified. All arguments are strings
}
