% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drugs_ndc.R
\name{get_drug_ndc}
\alias{get_drug_ndc}
\title{Get a dataset with flags indicating if patient has a medication or not using NDC codes}
\usage{
get_drug_ndc(supermart, sandbox, cohort, drug_list_df, ndc_col,
  grouping_col, index_col, date_condition)
}
\arguments{
\item{supermart}{Name of the supermart where the data is stored. Eg. supermart_392}

\item{sandbox}{Name of the writable location usually sandbox where cohort is saved}

\item{cohort}{Name of the cohort with explorys_patient_id's.}
}
\description{
This function provides a dataframe with flags indicating if a patient has a medication or not. The flags are 1 which
indicate patient had the medication or 0 which indicates patient did not. A missing flag means that the patient did not
get any medication at all in the time frame specified. All arguments are strings
}
