% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drugs_string.R
\name{get_drugs_string}
\alias{get_drugs_string}
\title{Get a dataset with flags indicating if patient has a medication or not using generic drug names}
\usage{
get_drugs_string(supermart, sandbox, cohort, drug_list_df, grouping_col,
  drug_names_col, index_col, date_condition, extra_grouping_vars = "")
}
\arguments{
\item{supermart}{Name of the supermart where the data is stored. Eg. supermart_392}

\item{sandbox}{Name of the writable location usually sandbox where cohort is saved}

\item{cohort}{Name of the cohort with explorys_patient_id's.}

\item{drug_list_df}{Name of the dataframe containing the medications of interest. This dataframe must be available in local environment as well as
on the sandbox}

\item{grouping_col}{Name of the column that has the groups of medications of interest}

\item{drug_names_col}{Name of the column in the drug_list_df specified. This column has the drug names of interest}

\item{index_col}{Name of the column with the index dates}

\item{date_condition}{The sql format date condition. For example "DATE(diagnosis_date) BETWEEN '2019-01-01' AND '2020-01-01'"}

\item{extra_grouping_vars}{Character string of additional variables separated by comma to grouping variables}
}
\value{
A dataframe with indicators of medication usage
}
\description{
This function returns a dataframe with flags indicating if a patient has a medication or not. The flags are 1 which
indicate patient had the medication or 0 which indicates patient did not. A missing flag means that the patient did not
get any medication at all in the time frame specified. All arguments are strings
}
