% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_results.R
\name{table_results}
\alias{table_results}
\title{Easily get results presentable for excel
This function allows you to get results that is easily presented in excel table shells. Usually, one may want
results of counts of a list of diagnoses. The dataset that goes into this function has a list of variables that
flag the whether a person has the event or not. Recommended for use for datasets derived using function like `get_diagnosis` and
`get_drugs_ndc` and similar functions from the explorys package.}
\usage{
table_results(df, levels)
}
\arguments{
\item{df}{Datafram with the flaging variables (1 or 0). Usually dataframe from functions like `get_diagnosis` among others from explorys
package}
}
\description{
Easily get results presentable for excel
This function allows you to get results that is easily presented in excel table shells. Usually, one may want
results of counts of a list of diagnoses. The dataset that goes into this function has a list of variables that
flag the whether a person has the event or not. Recommended for use for datasets derived using function like `get_diagnosis` and
`get_drugs_ndc` and similar functions from the explorys package.
}
