% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{get_demographics}
\alias{get_demographics}
\title{Get demographic characteristics of patients in explorys database}
\usage{
get_demographics(supermart, sandbox, cohort)
}
\arguments{
\item{supermart}{Name of the supermart where the data is stored. Eg. supermart_392}

\item{sandbox}{Name of the writable location usually sandbox where cohort is saved}

\item{cohort}{Name of the cohort with explorys_patient_id's and index dates. Must be called index_dt}
}
\value{
A dataframe with flags of counts of demographic variables
}
\description{
This function takes in the cohort with ids and index date and outputs the demographic characteristics like gender, age and 
and groups, insurance, race/ethnicity. It assumes the cohort has variables called index_dt and explorys_patient_id. 
Always make sure those 2 variables exists with the exact same names.
The results are aggregated from the entire history of patients existence in the database. So for example if a patient reports 
Medicare as their insurance at some point in time and later reports Medicaid, they will be counted in both categories.
}
